refactor: improve(?) annotation processor structure
