+++ annotation-processor/src/main/java/me/bristermitten/mittenlib/annotations/config/ConfigClassBuilder.java
@@ -1,21 +1,23 @@
 package me.bristermitten.mittenlib.annotations.config;
 
-import com.google.gson.reflect.TypeToken;
 import com.squareup.javapoet.*;
 import me.bristermitten.mittenlib.annotations.util.ElementsFinder;
 import me.bristermitten.mittenlib.annotations.util.Stringify;
 import me.bristermitten.mittenlib.annotations.util.TypesUtil;
-import me.bristermitten.mittenlib.config.*;
+import me.bristermitten.mittenlib.config.Config;
+import me.bristermitten.mittenlib.config.Configuration;
+import me.bristermitten.mittenlib.config.GeneratedConfig;
+import me.bristermitten.mittenlib.config.Source;
 import me.bristermitten.mittenlib.config.generate.GenerateToString;
 import me.bristermitten.mittenlib.util.Result;
 import me.bristermitten.mittenlib.util.Strings;
-import org.jetbrains.annotations.Contract;
-import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import javax.inject.Inject;
-import javax.lang.model.element.*;
-import javax.lang.model.type.DeclaredType;
+import javax.lang.model.element.Element;
+import javax.lang.model.element.Modifier;
+import javax.lang.model.element.TypeElement;
+import javax.lang.model.element.VariableElement;
 import javax.lang.model.type.TypeKind;
 import javax.lang.model.type.TypeMirror;
 import javax.lang.model.util.Types;
@@ -47,6 +49,10 @@ public class ConfigClassBuilder {
 
     private final GeneratedTypeCache generatedTypeCache;
 
+    private final DeserializationCodeGenerator deserializationCodeGenerator;
+    private final ConstructorGenerator constructorGenerator;
+    private final AccessorGenerator accessorGenerator;
+
     @Inject
     ConfigClassBuilder(ElementsFinder elementsFinder,
                        Types types,
@@ -54,7 +60,11 @@ public class ConfigClassBuilder {
                        TypesUtil typesUtil,
                        ConfigurationClassNameGenerator classNameGenerator,
                        ToStringGenerator toStringGenerator,
-                       FieldClassNameGenerator fieldClassNameGenerator, GeneratedTypeCache generatedTypeCache) {
+                       FieldClassNameGenerator fieldClassNameGenerator,
+                       GeneratedTypeCache generatedTypeCache,
+                       DeserializationCodeGenerator deserializationCodeGenerator,
+                       ConstructorGenerator constructorGenerator,
+                       AccessorGenerator accessorGenerator) {
         this.elementsFinder = elementsFinder;
         this.types = types;
         this.methodNames = methodNames;
@@ -63,6 +73,9 @@ public class ConfigClassBuilder {
         this.toStringGenerator = toStringGenerator;
         this.fieldClassNameGenerator = fieldClassNameGenerator;
         this.generatedTypeCache = generatedTypeCache;
+        this.deserializationCodeGenerator = deserializationCodeGenerator;
+        this.constructorGenerator = constructorGenerator;
+        this.accessorGenerator = accessorGenerator;
     }
 
     private FieldSpec createFieldSpec(VariableElement element) {
@@ -101,26 +114,6 @@ public class ConfigClassBuilder {
         return typesUtil.getConfigClassName(typeMirror, source);
     }
 
-
-    private void createGetterMethod(TypeSpec.Builder typeSpecBuilder, VariableElement element, FieldSpec field) {
-        var safeName = getFieldAccessorName(element);
-
-        var builder = MethodSpec.methodBuilder(safeName)
-                .addModifiers(Modifier.PUBLIC)
-                .returns(field.type)
-                .addStatement("return " + field.name);
-
-        if (typesUtil.isNullable(element)) {
-            builder.addAnnotation(Nullable.class);
-        } else {
-            builder.addAnnotation(NotNull.class);
-        }
-
-        builder.addAnnotation(AnnotationSpec.builder(Contract.class)
-                .addMember("pure", CodeBlock.of("true")).build());
-        typeSpecBuilder.addMethod(builder.build());
-    }
-
     private TypeSpec createConfigClass(TypeElement classType,
                                        List<VariableElement> variableElements) {
         final ClassName className =
@@ -141,12 +134,6 @@ public class ConfigClassBuilder {
         return superClass;
     }
 
-    private String getSuperFieldName(TypeMirror superClass) {
-        var configName = typesUtil.getConfigClassName(superClass);
-
-        return "parent" + Strings.capitalize(typesUtil.getSimpleName(configName));
-    }
-
     private TypeSpec createConfigClass(TypeElement classType,
                                        List<VariableElement> variableElements,
                                        ClassName className) {
@@ -165,7 +152,7 @@ public class ConfigClassBuilder {
             var superclassName = getConfigClassName(superClass, classType);
             typeSpecBuilder.superclass(superclassName);
 
-            var superParamName = getSuperFieldName(superClass);
+            var superParamName = constructorGenerator.getSuperFieldName(superClass);
             typeSpecBuilder.addField(FieldSpec.builder(superclassName, superParamName)
                     .addModifiers(Modifier.PRIVATE, Modifier.FINAL)
                     .build()
@@ -185,7 +172,7 @@ public class ConfigClassBuilder {
         createDeserializeMethod(typeSpecBuilder, classType, className, variableElements);
 
         // Generate getter methods
-        fieldSpecs.forEach((elem, field) -> createGetterMethod(typeSpecBuilder, elem, field));
+        fieldSpecs.forEach((elem, field) -> accessorGenerator.createGetterMethod(typeSpecBuilder, elem, field));
 
         // Generate copy setter methods
         fieldSpecs.values().forEach(field -> {
@@ -200,7 +187,7 @@ public class ConfigClassBuilder {
 
             if (superClass != null) {
                 var joiner = new StringJoiner(",")
-                        .add(getSuperFieldName(superClass));
+                        .add(constructorGenerator.getSuperFieldName(superClass));
                 if (!constructorParams.isEmpty()) {
                     joiner.add(constructorParams);
                 }
@@ -248,41 +235,15 @@ public class ConfigClassBuilder {
                                        Collection<FieldSpec> fieldSpecs,
                                        TypeSpec.Builder typeSpecBuilder,
                                        @Nullable TypeMirror superclass) {
-        final MethodSpec.Builder constructorBuilder = MethodSpec.constructorBuilder()
-                .addModifiers(Modifier.PUBLIC);
-
-        if (superclass != null) {
-            var superclassName = getConfigClassName(superclass, variableElements.stream().findFirst().map(VariableElement::getEnclosingElement).orElse(null));
-            var superParamName = getSuperFieldName(superclass);
-            var parameter = ParameterSpec.builder(superclassName, superParamName)
-                    .addModifiers(Modifier.FINAL)
-                    .build();
-            constructorBuilder.addParameter(parameter);
-
-
-            var superElements = elementsFinder.getApplicableVariableElements(superclass);
-            List<String> collect = superElements.stream()
-                    .map(variableElement -> superParamName + "." + getFieldAccessorName(variableElement) + "()")
-                    .toList();
-
-            if (getDTOSuperclass((TypeElement) types.asElement(superclass)) != null) {
-                var newCollect = new ArrayList<>(collect);
-                newCollect.addFirst(superParamName);
-                collect = newCollect;
-            }
-
-            constructorBuilder.addStatement("super($L)", String.join(", ", collect));
-
-            constructorBuilder.addStatement("this.$L = $L", superParamName, superParamName);
-        }
-
-        constructorBuilder.addParameters(variableElements.stream()
-                .map(this::createParameterSpec)
-                .toList());
-
-        fieldSpecs.forEach(field ->
-                constructorBuilder.addStatement("this.$N = $N", field, field.name));
-        typeSpecBuilder.addMethod(constructorBuilder.build());
+        constructorGenerator.addAllArgsConstructor(
+                variableElements,
+                fieldSpecs,
+                typeSpecBuilder,
+                superclass,
+                typeMirror -> getConfigClassName(typeMirror, variableElements.stream().findFirst().map(VariableElement::getEnclosingElement).orElse(null)),
+                this::getDTOSuperclass,
+                this::getFieldAccessorName
+        );
     }
 
     private void createConfigurationField(TypeElement classType, ClassName className, TypeSpec.Builder
@@ -306,162 +267,29 @@ public class ConfigClassBuilder {
     }
 
     private MethodSpec createDeserializeMethodFor(TypeElement dtoType, VariableElement element) {
-        TypeMirror typeMirror = element.asType();
-        final TypeName elementType = getConfigClassName(typeMirror, dtoType);
-        final Name variableName = element.getSimpleName();
-        final TypeMirror boxedType = typesUtil.getBoxedType(typeMirror);
-        final TypeName boxedTypeName = getConfigClassName(boxedType, dtoType);
-
-        final MethodSpec.Builder builder = MethodSpec.methodBuilder(DESERIALIZE_METHOD_PREFIX + Strings.capitalize(variableName.toString()))
-                .addModifiers(Modifier.PRIVATE, Modifier.STATIC)
-                .returns(ParameterizedTypeName.get(RESULT_CLASS_NAME, boxedTypeName))
-                .addParameter(ParameterSpec.builder(DeserializationContext.class, "context").build())
-                .addParameter(ParameterSpec.builder(ClassName.get(dtoType), "dao").build());
-
-        builder.addStatement("$T $$data = context.getData()", MAP_STRING_OBJ_NAME);
-        builder.addStatement("$T $L", elementType, variableName);
-        final String fromMapName = variableName + "FromMap";
-        final String key = fieldClassNameGenerator.getConfigFieldName(element);
-        var defaultName = element.getSimpleName();
-        builder.addStatement("Object $L = $$data.getOrDefault($S, dao.$L)", fromMapName, key, defaultName);
-
-        if (typesUtil.isNullable(element)) {
-            // Short circuit the null rather than trying any deserialization
-            builder.beginControlFlow("if ($L == null)", fromMapName);
-            builder.addStatement("return $T.ok(null)", Result.class);
-            builder.endControlFlow();
-        } else {
-            builder.beginControlFlow("if ($L == null)", fromMapName);
-            builder.addStatement("return $T.fail($T.throwNotFound($S, $S, $T.class, $L))",
-                    Result.class, ConfigMapLoader.class, variableName, element, dtoType, fromMapName);
-            builder.endControlFlow();
-        }
-
-        if (!(elementType instanceof ParameterizedTypeName)) {
-            /*
-             Construct a simple check that does
-               if (fromMap instanceof X) return fromMap; NOSONAR this is not code you stupid program
-             Useful when the type is a primitive or String
-             This is only safe to do with non-parameterized types, type erasure and all
-            */
-            final TypeMirror safeType = typesUtil.getSafeType(typeMirror);
-            final TypeName safeTypeName = getConfigClassName(safeType, dtoType);
-            builder.beginControlFlow("if ($L instanceof $T)", fromMapName, safeTypeName);
-            builder.addStatement("return $T.ok(($T) $L)", Result.class, elementType, fromMapName);
-            builder.endControlFlow();
-        } else if (typeMirror instanceof DeclaredType declaredType) {
-            /*
-             This is really cursed, but it's the only real way
-             When the type is a List<T> or Map<_, T> then we need to first load it as a C<Map<String, Object>> then
-             apply the deserialize function to each element. Otherwise, gson would try to deserialize it without
-             using the generated deserialization method and produce inconsistent results.
-
-             However, we can't just blindly do this for every type - there's no way of knowing how to convert a
-             Blah<A> into a Blah<B> without some knowledge of the underlying structure.
-             Map and List are the most common collection types, but this could really use some extensibility.
-            */
-            String canonicalName = types.erasure(typeMirror).toString();
-            if (canonicalName.equals("java.util.List")) {
-                var listType = declaredType.getTypeArguments().getFirst();
-                if (isConfigType(listType)) {
-                    TypeName listTypeName = getConfigClassName(listType, null);
-                    builder.addStatement("return $T.deserializeList($L, context, $T::$L)", CollectionsUtils.class,
-                            fromMapName,
-                            listTypeName,
-                            getDeserializeMethodName(listTypeName));
-                    return builder.build();
-                }
-            }
-            if (canonicalName.equals("java.util.Map")) {
-                var mapType = declaredType.getTypeArguments().get(1);
-                var keyType = declaredType.getTypeArguments().get(0);
-
-                if (isConfigType(mapType)) {
-                    TypeName mapTypeName = getConfigClassName(mapType, null);
-                    builder.addStatement("return $T.deserializeMap($T.class, $L, context, $T::$L)",
-                            CollectionsUtils.class,
-                            getConfigClassName(typesUtil.getSafeType(keyType), null),
-                            fromMapName,
-                            mapTypeName,
-                            getDeserializeMethodName(mapTypeName));
-                    return builder.build();
-                }
-            }
-        }
-
-        if (isConfigType(typeMirror)) {
-            builder.beginControlFlow("if ($L instanceof $T)", fromMapName, Map.class);
-            builder.addStatement("$1T mapData = ($1T) $2L", MAP_STRING_OBJ_NAME, fromMapName);
-            builder.addStatement("return $T.$L(context.withData(mapData))", elementType, getDeserializeMethodName(elementType));
-            builder.endControlFlow();
-        }
-
-
-        // If no shortcuts work, pass it to the context and do some dynamic-ish deserialization
-        builder.addStatement("return context.getMapper().map($N, new $T<$T>(){})", fromMapName, TypeToken.class, boxedTypeName);
-        return builder.build();
+        return deserializationCodeGenerator.createDeserializeMethodFor(dtoType, element);
     }
 
     private boolean isConfigType(TypeMirror mirror) {
-        return mirror instanceof DeclaredType declaredType &&
-                declaredType.asElement().getAnnotation(Config.class) != null;
+        return deserializationCodeGenerator.isConfigType(mirror);
     }
 
 
     private String getDeserializeMethodName(TypeName name) {
-        if (name instanceof ClassName cn) {
-            return DESERIALIZE_METHOD_PREFIX + cn.simpleName();
-        }
-        return DESERIALIZE_METHOD_PREFIX + name;
+        return deserializationCodeGenerator.getDeserializeMethodName(name);
     }
 
     private void createDeserializeMethod(TypeSpec.Builder typeSpecBuilder,
                                          TypeElement dtoType,
                                          ClassName className,
                                          List<VariableElement> variableElements) {
-
-        final MethodSpec.Builder builder = MethodSpec.methodBuilder(getDeserializeMethodName(className))
-                .addModifiers(Modifier.PUBLIC, Modifier.STATIC)
-                .returns(ParameterizedTypeName.get(RESULT_CLASS_NAME, className))
-                .addParameter(ParameterSpec.builder(DeserializationContext.class, "context").addModifiers(Modifier.FINAL).build());
-
-        builder.addStatement("$1T dto = new $1T()", (dtoType.asType()));
-
-        final List<MethodSpec> deserializeMethods = variableElements.stream()
-                .map(variableElement -> createDeserializeMethodFor(dtoType, variableElement))
-                .toList();
-
-        deserializeMethods.forEach(typeSpecBuilder::addMethod);
-
-        final CodeBlock.Builder expressionBuilder = CodeBlock.builder();
-
-        expressionBuilder.add("return ");
-        int i = 0;
-        var superClass = getDTOSuperclass(dtoType);
-
-        // Add the superclass deserialization first, if it exists
-        if (superClass != null) {
-            var superConfigName = getConfigClassName(superClass, dtoType);
-            expressionBuilder.add("$T.$L", superConfigName, getDeserializeMethodName(superConfigName));
-            expressionBuilder.add("(context).flatMap(var$L -> \n", i++);
-        }
-        for (MethodSpec deserializeMethod : deserializeMethods) {
-            expressionBuilder.add("$N(context, dto).flatMap(var$L -> \n", deserializeMethod, i++);
-        }
-        expressionBuilder.add("$T.ok(new $T(", Result.class, className);
-        for (int i1 = 0; i1 < i; i1++) {
-            expressionBuilder.add("var$L", i1);
-            if (i1 != i - 1) {
-                expressionBuilder.add(", ");
-            }
-        }
-        expressionBuilder.add("))"); // Close ok and new parens
-        expressionBuilder.add(")".repeat(Math.max(0, i))); // close all the flatMap parens
-
-
-        builder.addStatement(expressionBuilder.build());
-
-        typeSpecBuilder.addMethod(builder.build());
+        deserializationCodeGenerator.createDeserializeMethod(
+                typeSpecBuilder,
+                dtoType,
+                className,
+                variableElements,
+                this::getDTOSuperclass
+        );
     }
 
 }
